server/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AWS
boto3==1.29.7
botocore==1.32.7

# SFTP
paramiko==3.3.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS and Security
python-multipart==0.0.6
itsdangerous==2.1.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging
python-json-logger==2.0.7

# Redis (optional, for caching)
redis==5.0.1

# Email
aiosmtplib==3.0.1

# System monitoring
psutil==5.9.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2